legacy/requirements.txt
requests
openpyxl
bs4
pandas
xlrd
lxml
python-dateutil
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
twilio
phonenumbers
peopledatalabs
visidata
geopy
paramiko
doltpy
sib_api_v3_sdk
